Added supported perl versions on travis to linter.
This is the second try after pull#33.

I now added the supported perl versions in the correct place.
Also, I found not only the problem is that perl 5.8 is supported on
Travis and is rejected by the linter (although I understand it is
not /fully/ supported, or at the least, if it breaks not many people
will care), but also perl 5.18, which is the latest stable perl, and
perl 5.19 ARE available on travis but are rejected by the linter.

So I made sure these are now accepted.